+++ buildSrc/src/main/kotlin/CONST.kt
@@ -1,3 +1,3 @@
-const val JAVA_TARGET_V_int = 11
+const val JAVA_TARGET_V_int = 17
 const val JAVA_TARGET_V = "$JAVA_TARGET_V_int"