+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/internal/configuration/BotConfiguration.kt
@@ -4,6 +4,7 @@ import eu.vendeli.tgbot.implementations.ClassManagerImpl
 import eu.vendeli.tgbot.implementations.InputListenerMapImpl
 import eu.vendeli.tgbot.interfaces.ctx.ClassManager
 import eu.vendeli.tgbot.interfaces.ctx.InputListener
+import eu.vendeli.tgbot.types.internal.ExceptionHandlingStrategy
 
 /**
  * The class containing the bot configuration.
@@ -14,7 +15,7 @@ import eu.vendeli.tgbot.interfaces.ctx.InputListener
  * @property inputListener Input handling class instance.
  * @property classManager The manager that will be used to get classes.
  * @property inputAutoRemoval A flag that regulates the auto-deletion of the input point during processing.
- * @property catchExceptions On true > catch raised during processing exceptions and wrap it to caughtExceptions.
+ * @property exceptionHandlingStrategy Exception handling strategy.
  * @property throwExOnActionsFailure Throw exception when the action (any bot request) ends with failure.
  */
 data class BotConfiguration(
@@ -24,7 +25,7 @@ data class BotConfiguration(
     var inputListener: InputListener = InputListenerMapImpl(),
     var classManager: ClassManager = ClassManagerImpl(),
     var inputAutoRemoval: Boolean = true,
-    var catchExceptions: Boolean = true,
+    var exceptionHandlingStrategy: ExceptionHandlingStrategy = ExceptionHandlingStrategy.CollectToChannel,
     var throwExOnActionsFailure: Boolean = false,
     internal var rateLimiter: RateLimiterConfiguration = RateLimiterConfiguration(),
     internal var httpClient: HttpConfiguration = HttpConfiguration(),